Item Page as per .docx document
most of item page edits requested by Karolina and Hanka have been implemented, with the notable exception of language qualifier selection of fields, as this feature is not implemented in DSpace yet. Some of the former look of simple item page was left in the code, most notably journal metadata in left collumn and citations identifier in right.