add user manual link to footer
add link to PDF copy of user manual to footer. Added via simple <a href=> where the location is the same as is in the wiki.techlib page. This has the effect that the user needs to log into techlib account first in order to see the manual, which is somehow desirable.